# Performance: Optimize ASCII spans in `seems_utf8()`.