@dmsnell 2025-08-12
Closing in favor of #9317 where `seems_utf8()` is deprecated.